feat: add native TS forge evaluator (not all language features supported yet)
Earlier, the way to evaluate things via forge within Sterling was to call some asynchronous hooks and wait for some results to be set in some global context, and once they were found there they could be used. However, having this level of asynchrony was leading to a lot of complexity in the implementation of the visualization generation, particularly in cases of conditional and nested expressions. Therefore, we decided that it would be nice to have a native TypeScript evaluator for Forge expressions (which would be synchronous). This is the start of that work -- it is an adapted version of a (Forge parser)[https://github.com/sidprasad/forge-antlr/tree/main] developed by Siddhartha Prasad, with some changes to allow it to work on individual expressions instead of a whole Forge spec, and a custom evaluator for those expressions